requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.12
frozenlist==1.6.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.0
greenlet==3.2.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
multidict==6.4.3
openai==1.75.0
packaging==24.2
playwright==1.51.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyee==12.1.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
readmeai==0.6.1
realtime==2.4.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
structlog==24.4.0
supabase==2.15.0
supafunc==0.9.4
tenacity==8.5.0
tiktoken==0.4.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==14.2
yarl==1.20.0